created a carousel for complete and incomplete projects. crud box working in projects page by rendering correct component onclick
