+++ frontend/src/components/Carousel/carousel.css
@@ -1,21 +1,20 @@
 #carousel-container {
   display: flex;
   justify-content: center;
-  margin-top: 50px;
   align-items: center;
+  width: 100%;
+  height: 100%;
+}
+#image-container {
+  display: flex;
+  margin: 0;
+  justify-content: center;
 }
 
 #carousel-image {
-  object-fit: contain;
-  max-width: 100%;
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.6);
   cursor: pointer;
-}
-
-#image-container {
-  display: flex;
-  margin: 0 20px;
-  justify-content: center;
+  object-fit: contain;
 }
 
 .carousel-content {
@@ -28,6 +27,8 @@
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.3);
   padding: 5px 15px;
   border-radius: 3px;
+  height: 20px;
+  margin: 0px 20px;
 }
 
 .carousel-buttons:hover {
