--- frontend/src/components/Projects/projects.css
@@ -1,38 +1,0 @@
-.projects-container {
-  display: flex;
-  flex-direction: column;
-  margin: 0 100px;
-  margin-top: 50px;
-}
-
-#blueprint-project-title {
-  cursor: default;
-}
-
-.project {
-  cursor: default;
-}
-
-.project-titles {
-  border-bottom: 1px solid black;
-  padding-bottom: 15px;
-  padding-left: 15px;
-  cursor: default;
-}
-
-.project-carousel-container {
-  margin-left: 50px;
-  height: 200px;
-  width: 300px;
-  border: 1px solid black;
-}
-
-.completed-projects {
-  width: 100%;
-  /* height: 100%; */
-  cursor: pointer;
-}
-
-#create-project-container {
-  width: 240px;
-}
