+++ frontend/src/index.css
@@ -2,15 +2,15 @@
 html {
   font-family: "Roboto";
 }
+#root {
+  padding: 0;
+  margin: 0;
+}
 
 body {
   padding: 0;
   margin: 0;
-  overflow: hidden;
+  overflow-x: hidden;
   height: 100vh;
-}
-
-#root {
-  padding: 0;
-  margin: 0;
+  position: relative;
 }
